+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - (BREAKING) Store enum values as string (impact ordering) - Algo (#1040)
+- (BREAKING) Add mandatory name field to compute plan (#1027)
 
 ## [0.13.0] 2022-05-09